+++ .github/workflows/ci.yml
@@ -38,3 +38,19 @@ jobs:
         uses: mfinelli/setup-shfmt@v3
       - name: Run shfmt
         run: shfmt -d .
+
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ env.PYTHON_VERSION }}
+
+      - name: Install dependencies
+        run: make install
+
+      - name: Run tests
+        run: make test
