+++ requirements/test.in
@@ -1,0 +1,1 @@
+pytest
