+++ test/assets/hash_function.py
@@ -1,0 +1,2 @@
+def get_hash() -> str:
+    return "plugin_id_fn_123"
