+++ test/assets/improper_function.py
@@ -1,0 +1,2 @@
+def sample_improper_function(*args, **kwargs):
+    pass
