+++ test/assets/simple_hash_lambda.py
@@ -1,0 +1,1 @@
+hash_lambda_fn = lambda: "plugin_id_hash_123"  # noqa: E731
