+++ test/assets/simple_hash_value.py
@@ -1,0 +1,1 @@
+hash_value = "plugin_id_123"
