+++ unstructured_platform_plugins/etl_uvicorn/json_schema.py
@@ -23,6 +23,8 @@ typed_map_reverse: dict[str, Type] = {v: k for k, v in types_map.items()}
 
 
 def is_optional(val: Any) -> bool:
+    if annotation := getattr(val, "annotation", None):
+        return is_optional(annotation)
     if origin := getattr(val, "__origin__", None):  # noqa: SIM102
         if origin is typing.Union:
             types = val.__args__
@@ -94,7 +96,7 @@ def pydantic_base_model_to_json_schema(model: Type[BaseModel]) -> dict:
     for name, f in fs.items():
         t = f.annotation
         f_resp = to_json_schema(t)
-        if f.default and f.default != PydanticUndefined:
+        if f.default != PydanticUndefined:
             f_resp["default"] = f.default
         properties[name] = f_resp
         if not is_optional(t):
@@ -175,8 +177,7 @@ def parameters_to_json_schema(parameters: list[Parameter]) -> dict:
     required_fields = []
     for p in parameters:
         schema = to_json_schema(val=p)
-        required = not is_optional(p)
-        if required:
+        if not is_optional(p):
             required_fields.append(p.name)
         resp[p.name] = schema
     if required_fields:
@@ -194,7 +195,7 @@ def run_output_checks(return_annotation: Any):
         return
     if is_dataclass(return_annotation):
         return
-    if issubclass(return_annotation, BaseModel):
+    if inspect.isclass(return_annotation) and issubclass(return_annotation, BaseModel):
         return
     if return_annotation is type(None):
         return
